% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateCelltype.r
\name{AnnotateCelltype}
\alias{AnnotateCelltype}
\title{Cell type annotation}
\usage{
AnnotateCelltype(
  scRNA,
  model = "claude-3-5-sonnet-20240620",
  tissuename = "unknow",
  api_key = "",
  base_url = ""
)
}
\arguments{
\item{scRNA}{: Seurat object}

\item{model}{: The model name you want to use}

\item{tissuename}{: Cell derived tissue}

\item{api_key}{: Key for API interface}

\item{base_url}{: API interface address}
}
\value{
: Annotate the cell types of the Seurat object.
}
\description{
Cell type annotation
}
