% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotationCelltype.r
\name{AnnotationCelltype}
\alias{AnnotationCelltype}
\title{Cell type annotation}
\arguments{
\item{scRNA}{: Seurat object}

\item{model}{: The model name you want to use}

\item{tissuename}{: Cell derived tissue}

\item{api_key}{: Key for API interface}

\item{base_url}{: API interface address}
}
\value{
: Annotate the cell types of the Seurat object.
}
\description{
Cell type annotation
}
