% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellType.r
\name{GetCellType}
\alias{GetCellType}
\title{Enter the markergenes list to obtain the corresponding cell type}
\usage{
GetCellType(
  markergenes,
  tissuename,
  model = "claude-3-5-sonnet-20240620",
  api_key = "",
  base_url = "",
  temperature = 0.1
)
}
\arguments{
\item{tissuename}{: Cell derived tissue}

\item{model}{: The model name you want to use}

\item{api_key}{: Key for API interface}

\item{base_url}{: API interface address}

\item{markergenes:}{A list of genes in string form, separated by rows, separated by genes within the same cluster}
}
\value{
The List of cell types
}
\description{
Enter the markergenes list to obtain the corresponding cell type
}
