% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sc_KPPC_aLiver_2}
\alias{sc_KPPC_aLiver_2}
\title{肝脏组织单细胞RNA-seq数据示例}
\description{
这是一个来自肝脏组织的单细胞RNA-seq数据集，以Seurat对象的形式存储。
数据来源于KPPC（Kras G12D, p53 R172H, Pdx-1-Cre）小鼠模型的肝脏组织。
}
\format{
An object of class \code{Seurat} with the following components:
\describe{
  \item{assays}{包含RNA表达数据}
  \item{meta.data}{包含细胞的元数据信息，如seurat_clusters等}
  \item{reductions}{包含降维结果，如UMAP、PCA等}
}
}
\source{
内部生成的示例数据
}
\examples{
# 加载示例数据
library('AIcelltype')
data(sc_KPPC_aLiver_2)

# 查看数据基本信息
dim(sc_KPPC_aLiver_2)
head(sc_KPPC_aLiver_2@meta.data)

# 也可以使用便捷函数加载数据
load_example_data()
}
\seealso{
\code{\link{load_example_data}}, \code{\link{ai_celltype}}
}