% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellType.r
\docType{data}
\name{url}
\alias{url}
\title{Enter the markergenes list to obtain the corresponding cell type}
\format{
An object of class \code{character} of length 1.
}
\usage{
url
}
\arguments{
\item{markergenes:}{A list of genes in string form, separated by rows, separated by genes within the same cluster}

\item{model}{: The model name you want to use}

\item{tissuename}{: Cell derived tissue}

\item{api_key}{: Key for API interface}

\item{base_url}{: API interface address}
}
\value{
The List of cell types
}
\description{
Enter the markergenes list to obtain the corresponding cell type
}
\keyword{datasets}
